Renamed and pylinted release_info.py
