+++ tcsfw/release_info.py
@@ -1,3 +1,5 @@
+"""Release information"""
+
 import datetime
 from typing import Optional
 
@@ -5,6 +7,7 @@ from tcsfw.property import PropertyKey
 
 
 class ReleaseInfo:
+    """Release information for software, firmware, etc. (NOTE: Not really used now)"""
     def __init__(self, sw_name: str):
         self.sw_name = sw_name
         self.latest_release: Optional[datetime.datetime] = None
@@ -14,10 +17,12 @@ class ReleaseInfo:
 
     @classmethod
     def parse_time(cls, date: Optional[str]) -> Optional[datetime.datetime]:
+        """Parse date string to object"""
         return datetime.datetime.strptime(date, '%Y-%m-%d') if date else None
 
     @classmethod
     def print_time(cls, date: Optional[datetime.datetime]) -> str:
+        """Print date object as string"""
         return date.strftime('%Y-%m-%d') if date else ""
 
     def __repr__(self):
